Step5 로또(수동) (#580)
* docs: 5단계 수동 요구사항 추가

* feat: 5단계 수동 구매

* refactor: LottoCount,LottoGame 생성 변경

* refactor: 매직넘버 제거 및 빈 라인 추가

* refactor: Money의 value 변경(int -> long)

+++ src/main/java/lotto/domain/Money.java
@@ -4,20 +4,20 @@ import java.util.Objects;
 
 public class Money {
     private static final int MIN = 0;
-    private final int value;
+    private final long value;
 
-    private Money(int value) {
+    private Money(long value) {
         validate(value);
         this.value = value;
     }
 
-    private void validate(int value) {
+    private void validate(long value) {
         if (value < MIN) {
             throw new IllegalArgumentException(String.format("돈은 %d보다 작을 수 없습니다.", MIN));
         }
     }
 
-    public static Money from(int money) {
+    public static Money from(long money) {
         return new Money(money);
     }
